# notebook-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NotebookAPI",
        "description": "L5 Swagger OpenApi description",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/notebook": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Получить список записей в записной книжке",
                "description": "Возвращает список записей в записной книжке",
                "operationId": "getNotebookEntries",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NotebookEntry"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/notebook": {
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Создать новую запись в записной книжке",
                "description": "Создает новую запись в записной книжке",
                "operationId": "createNotebookEntry",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NotebookEntry"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Запись успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotebookEntry"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Некорректные данные"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/notebook/{id}": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Получить информацию о записи в записной книжке",
                "description": "Возвращает информацию о записи в записной книжке по указанному ID",
                "operationId": "getNotebookEntryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID записи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotebookEntry"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Запись не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Запись не найдена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Обновить информацию о записи в записной книжке",
                "description": "Обновляет информацию о записи в записной книжке по указанному ID",
                "operationId": "updateNotebookEntry",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID записи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NotebookEntry"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Запись успешно обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotebookEntry"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Запись не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Запись не найдена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Удалить запись из записной книжки",
                "description": "Удаляет запись из записной книжки по указанному ID",
                "operationId": "deleteNotebookEntry",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID записи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Запись успешно удалена"
                    },
                    "404": {
                        "description": "Запись не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Запись не найдена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "NotebookEntry": {
                "required": [
                    "full_name",
                    "company",
                    "phone",
                    "email",
                    "birthdate",
                    "photo"
                ],
                "properties": {
                    "full_name": {
                        "type": "string",
                        "example": "FirstName SecondName"
                    },
                    "company": {
                        "type": "string",
                        "example": "ABC Corp"
                    },
                    "phone": {
                        "type": "string",
                        "example": "123-456-7890"
                    },
                    "email": {
                        "type": "string",
                        "example": "test@test.com"
                    },
                    "birthdate": {
                        "type": "string",
                        "format": "date",
                        "example": "2000-01-01"
                    },
                    "photo": {
                        "type": "string",
                        "format": "url",
                        "example": "https://test.com/photo.jpg"
                    }
                },
                "type": "object"
            }
        }
    }
}